refactor(data): 패키지 수정
