+++ app/src/main/java/woowacourse/shopping/feature/cart/CartActivity.kt
@@ -8,9 +8,9 @@ import android.view.View
 import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
 import woowacourse.shopping.R
-import woowacourse.shopping.data.CartCache
-import woowacourse.shopping.data.CartRemoteRepositoryImpl
 import woowacourse.shopping.data.TokenSharedPreference
+import woowacourse.shopping.data.cart.CartCache
+import woowacourse.shopping.data.cart.CartRemoteRepositoryImpl
 import woowacourse.shopping.data.service.CartRemoteService
 import woowacourse.shopping.databinding.ActivityCartBinding
 import woowacourse.shopping.feature.order.OrderActivity
