+++ app/src/main/java/woowacourse/shopping/feature/detail/DetailActivity.kt
@@ -11,9 +11,9 @@ import androidx.appcompat.app.AlertDialog
 import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
 import woowacourse.shopping.R
-import woowacourse.shopping.data.CartCache
-import woowacourse.shopping.data.CartRemoteRepositoryImpl
 import woowacourse.shopping.data.TokenSharedPreference
+import woowacourse.shopping.data.cart.CartCache
+import woowacourse.shopping.data.cart.CartRemoteRepositoryImpl
 import woowacourse.shopping.data.service.CartRemoteService
 import woowacourse.shopping.databinding.ActivityDetailBinding
 import woowacourse.shopping.databinding.DialogSelectCountBinding
